For some reason which I have not figured out yet `dmake` got much slower. Since it is a fixed dependency in CMake builds it slows things down quite a bit. These changes greatly improve the run-time of it without changing the output.

We still show figure out when it got slow to fix the actual root cause.

Before:
```
Benchmark 1: ./dmake
  Time (mean ± σ):     12.244 s ±  0.727 s    [User: 0.571 s, System: 11.501 s]
  Range (min … max):   11.643 s … 13.440 s    5 runs
```

After:
```
Benchmark 1: ./dmake
  Time (mean ± σ):      1.081 s ±  0.101 s    [User: 0.162 s, System: 0.911 s]
  Range (min … max):    0.989 s …  1.251 s    5 runs
```

This was tested using GCC 12 on WSL1 Kali Linux.